Stop hardcoding ~/.rlw as installation path, instead use the cloned source folder as prefix
*sigh* I've broken enough things so that the script requires a git clone somewhere to work.

This hasn't been tested yet, but the script should be able to find the folder where it
resides and put all its files there. This means that $WINEPREFIX and any other temporary files
will reside in a subdirectory of the source folder, and that "installation" just registers
the .desktop file without having to do any copying or git cloning itself.

The idea behind these changes is to make the script to be more portable. The user gets to pick
where the script and its files reside, without complicating the installation instructions
too much.

More tests will be done when I get to my Linux box.

TODO: Update the .desktop file. Maybe a dynamic .desktop generator is needed? Or should we
add rlw.sh to the user's $PATH somehow?